pyproject.toml
[project]
name = "mcp-gpt-response-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
]
